# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/candidates": {
            "get": {
                "description": "Get all candidates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Get Candidates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "bioLink": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    },
                                    "imageLink": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "percentage": {
                                        "type": "string"
                                    },
                                    "policy": {
                                        "type": "string"
                                    },
                                    "votedCount": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Create Candidate",
                "parameters": [
                    {
                        "description": "candidate detail",
                        "name": "candidate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    }
                }
            }
        },
        "/candidates/{id}": {
            "get": {
                "description": "Get selected candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Get Candidate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    }
                }
            },
            "put": {
                "description": "Update candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Update Candidate",
                "parameters": [
                    {
                        "description": "candidate detail",
                        "name": "candidate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Candidate"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Delete Candidate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiDeleteCandidateHandlerResponse"
                        }
                    }
                }
            }
        },
        "/election/export": {
            "get": {
                "description": "Export result as CSV",
                "tags": [
                    "election"
                ],
                "summary": "Export Result",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/election/result": {
            "get": {
                "description": "Get results from every candidates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "election"
                ],
                "summary": "Get Result",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "bioLink": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    },
                                    "imageLink": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "percentage": {
                                        "type": "string"
                                    },
                                    "policy": {
                                        "type": "string"
                                    },
                                    "votedCount": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/election/toggle": {
            "post": {
                "description": "Toggle vote to open or close",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "election"
                ],
                "summary": "Toggle Election",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiPostToggleResponse"
                        }
                    }
                }
            }
        },
        "/vote": {
            "post": {
                "description": "Vote candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vote"
                ],
                "summary": "Vote Candidate",
                "parameters": [
                    {
                        "description": "Vote Input",
                        "name": "voteInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vote.VoteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiPostVoteResponse"
                        }
                    }
                }
            }
        },
        "/vote/status": {
            "post": {
                "description": "Check vote status if user have the right to vote",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vote"
                ],
                "summary": "Check Vote Status",
                "parameters": [
                    {
                        "description": "national Id",
                        "name": "nationalId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vote.VoteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiPostToggleResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ApiDeleteCandidateHandlerResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.ApiPostToggleResponse": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.ApiPostVoteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.Candidate": {
            "type": "object",
            "properties": {
                "bioLink": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "policy": {
                    "type": "string"
                },
                "votedCount": {
                    "type": "integer"
                }
            }
        },
        "vote.VoteInput": {
            "type": "object",
            "properties": {
                "candidateId": {
                    "type": "integer"
                },
                "nationalId": {
                    "type": "string"
                }
            }
        }
    }
}